refactor(events): split GameEvent Display impl into display.rs
