+++ game/src/events/mod.rs
@@ -1,3 +1,4 @@
+pub mod display;
 pub mod types;
 pub use types::GameEvent;
